* Resolved "signed/unsigned" Level 3 warnings.

+++ Trans3_2/vc/trans3/common/mainfile.cpp
@@ -149,7 +149,7 @@ bool tagMainFile::open(const STRING fileName)
 
 	short len;
 	file >> len;
-	unsigned int i;
+	short i;
 	for (i = 0; i <= len; i++)
 	{
 		short sUnused;
